% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{.rp_checkAccessToken}
\alias{.rp_checkAccessToken}
\title{Check & Auto-Renew Bearer Tokens (Internal)}
\usage{
.rp_checkAccessToken(printMsg = FALSE, mins = 120)
}
\arguments{
\item{printMsg}{= (bool) Should outcome messages be printed out? defaults to FALSE}

\item{mins}{= (int) The number of minutes that the bearer token will be valid for.}
}
\value{
Checks validity of Bearer Token & auto-updates if needed. Assigns the new tokens in 'rp' \code{environment} and 'rp_tokens.rds' file
}
\description{
Check & Auto-Renew Bearer Tokens (Internal)
}
\examples{
\dontrun{
  # For Internal Use Prior to Making API Requests
  .rp_checkAccessToken(printMsg=FALSE, mins=120)
}
}
