% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\docType{data}
\name{.rp_env}
\alias{.rp_env}
\title{temporary working environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.rp_env
}
\value{
An auto generated \code{environment} to store our tokens
}
\description{
temporary working environment
}
\examples{
\dontrun{
  .rp_env <- new.env(parent = emptyenv())
}
}
\keyword{datasets}
