% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{.rp_make_qte_payload}
\alias{.rp_make_qte_payload}
\title{Build Dynamic Payload For rp_getQuote (Internal)}
\usage{
.rp_make_qte_payload(symbols, types)
}
\arguments{
\item{symbols}{= (string) Equity/ETF/Option symbol(s)}

\item{types}{= (string) The product type (ex. 'EQUITY' or 'OPTION')}
}
\value{
Returns a \code{list} in the appropriate payload format in case the user needs multiple symbols for quotes
}
\description{
Build Dynamic Payload For rp_getQuote (Internal)
}
\examples{
\dontrun{
  # Create the correct quote payload for AAPL and a SPY 631 Call 8/15/25 Expiration
  .rp_make_qte_payload(symbols=c("AAPL","SPY250815C00631000"), types=c("EQUITY","OPTION"))
}
}
