% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{.rp_read_tokens}
\alias{.rp_read_tokens}
\title{Request token file (Internal)}
\usage{
.rp_read_tokens()
}
\value{
Requests your token file 'rp_tokens.rds' from working directory & assigns a working \code{environment} if it exists
}
\description{
Request token file (Internal)
}
\examples{
\dontrun{
  # For Internal Use (assigns tokens inside of the 'rp' environment)
  .rp_read_tokens()
}
}
