% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_links.R
\name{extract_links}
\alias{extract_links}
\title{Easy URL Extraction}
\usage{
extract_links(url_name, pattern = "", ignore.case = F)
}
\arguments{
\item{url_name}{str: A http/s URL that contains the URLs the user wishes to extract .}

\item{pattern}{str: The regex pattern used to filter for the extracted links.
Default is set to return all URLs}

\item{ignore.case}{bool: if \strong{TRUE}, the pattern matching is case
sensitive and if \strong{FALSE} case is ignored during matching.
Default is set to \strong{FALSE}}
}
\value{
A vector of \emph{n} strings containing all links that match the regex pattern
        in the pattern param.
}
\description{
Extracts all links in a given URL matching a specific regex pattern as a vector
of strings.
}
\examples{
## Extract all .xls files with Annual Report in the name from the RTT wait times data
## Make sure you have a decent internet connection!
rtt_url <- 'https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/'
Rpublic::extract_links(url_name = rtt_url,pattern='Annual-Report-spreadsheet|.xls')

}
