% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sheets.R
\name{extract_sheets}
\alias{extract_sheets}
\title{Sheet extraction from online excel workbook URLs}
\usage{
extract_sheets(files, pattern = "")
}
\arguments{
\item{files}{str: A URL to the .xls/.xlsx file of choice}

\item{pattern}{str: The regex pattern used to filter for the sheets in the .xlsx/.xls file}
}
\value{
A list of data frames of either/and .csv and .xlsx/.xls sheets matching specified regex patterns in a given Zip Archive URL
}
\description{
Extracts specific sheets as data.frames from excel urls based on a regex pattern
}
\examples{
## Extract all sheets matching the pattern string
## Make sure you have a decent internet connection!
nhsd_url <- 'https://www.england.nhs.uk/statistics/wp-content/uploads/sites/'
rtt_url <- '2/2024/07/Admitted-Commissioner-Mar24-XLSX-1914K-revised.xlsx'
Rpublic::extract_sheets(files = paste0(nhsd_url,rtt_url),pattern='Nation')

}
