% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_zipped.R
\name{extract_zipped}
\alias{extract_zipped}
\title{File Extraction from Online Zip Archives}
\usage{
extract_zipped(
  files,
  pattern,
  ignore.case = F,
  xlsx_flag = F,
  xlsx_sheet_pattern = "",
  ...
)
}
\arguments{
\item{files}{str: A URL to the zip archive of choice}

\item{pattern}{str: The regex pattern used to filter for the files in the Zip Archive}

\item{ignore.case}{bool: if \strong{FALSE} , the pattern matching is case sensitive and if \strong{FALSE} case is ignored during matching. Default is set to \strong{FALSE}}

\item{xlsx_flag}{bool: if \strong{FALSE} , the pattern ignores .xlsx/.xls files \strong{TRUE} The pattern matching will also extract Excel Sheets. Default is set to \strong{FALSE}}

\item{xlsx_sheet_pattern}{str: Regex pattern to match specific sheets in an .xlsx/.xls file. (See extract_excel_sheets) Default is set to return all sheets.}

\item{...}{Passes additional arguments to data.table::fread}
}
\value{
A list of data frames of either/and .csv and .xlsx/.xls sheets matching specified regex patterns in a given Zip Archive URL
}
\description{
Extracts specific files in from a Zip archive URL based on a set regex pattern
}
\examples{
## Extract all .csvs in
## Make sure you have a decent internet connection!
nhsd_url <- 'https://www.england.nhs.uk/statistics/wp-content/uploads/sites'
rtt_url <- '/2/2024/07/Full-CSV-data-file-Mar24-ZIP-3832K-revised.zip'
Rpublic::extract_zipped(files = paste0(nhsd_url,rtt_url),pattern='Annual-Report-spreadsheet|.xls')

}
