% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_bars}
\alias{rp_bars}
\title{Get Bars V2}
\usage{
rp_bars(
  symbol,
  period = "DAY",
  purchaseDate = NULL,
  type = "EQUITY",
  include_ah = FALSE
)
}
\arguments{
\item{symbol}{= The symbol to fetch bar data for: Ex. "SPY"}

\item{period}{= one of DAY, WEEK, MONTH, QUARTER, HALF_YEAR, YEAR, FIVE_YEARS, YTD, SINCE_PURCHASE}

\item{purchaseDate}{= Required for SINCE_PURCHASE. YYYY-MM-DD}

\item{type}{= Ticker Type: Ex. 'EQUITY','OPTION', 'CRYPTO', 'INDEX'}

\item{include_ah}{= Should after-hours data (if available) be included? Default is FALSE}
}
\value{
Fetch bar data for a given symbol and period and return as a \code{data.frame}.
}
\description{
Get Bars V2
}
\examples{
\dontrun{
 # Fetches Bars for the DAY - include after-hours data
 df <- rp_bars(symbol = "PLTR", period = "DAY", include_ah=TRUE)
 # Fetches Bars using SINCE_PURCHASE
 df <- rp_bars(symbol = "DASH", period = "SINCE_PURCHASE", 
 purchaseDate = "2026-05-14")
 # Fetches Bars for Options
 df <- rp_bars(symbol = "SPY260515C00730000", period = "WEEK", type="OPTION")
 # Get BTC Bars
 df <- rp_bars(symbol = "BTC", period = "DAY", type="CRYPTO")
}
}
