% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_cancel_order}
\alias{rp_cancel_order}
\title{Cancel Order}
\usage{
rp_cancel_order(accountId, orderId)
}
\arguments{
\item{accountId}{= Public Brokerage Account Number}

\item{orderId}{= The order ID}
}
\value{
Request order cancellation & return as a \code{data.frame}.
}
\description{
Cancel Order
}
\examples{
\dontrun{
 # Cancels Specific Order
 my_acc <- rp_getAccts()
 rp_cancel_order(accountId = my_acc$accountId, 
                 orderId = "c99be1dd-bb87-4f7a-803f-ec47226bf64e") 
}
}
