% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_fix_bars}
\alias{rp_fix_bars}
\title{Convert bars to data.frame}
\usage{
rp_fix_bars(ls2fix)
}
\arguments{
\item{ls2fix}{= List to convert}
}
\value{
Pass in list to convert it to a \code{data.frame}.
}
\description{
Convert bars to data.frame
}
\examples{
\dontrun{
 # Fetches Bars for the Week
 
 # Fetch Bars
 df <- rp_hist_bars("BTC", period = "WEEK", type = "CRYPTO", 
 aggregation = "FIFTEEN_MINUTES")
 # Extract the list you need, ex. regularMarket bars
 ls2fix <- df$regularMarket$bars
 # Pass it into function
 rp_fix_bars(ls2fix)
 
}
}
