% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_get_order}
\alias{rp_get_order}
\title{Get Order Details}
\usage{
rp_get_order(accountId, orderId)
}
\arguments{
\item{accountId}{= Public Brokerage Account Number}

\item{orderId}{= The order ID}
}
\value{
Retrieve order details & return as a \code{data.frame}.
}
\description{
Get Order Details
}
\examples{
\dontrun{
 # Fetches Specific Order
 my_acc <- rp_getAccts()
 rp_get_order(accountId = my_acc$accountId, 
              orderId = "c99be1dd-bb87-4f7a-803f-ec47226bf64e") 
}
}
