% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_getAccToken}
\alias{rp_getAccToken}
\title{Get New Access/Bearer Token From Secret Key}
\usage{
rp_getAccToken(exp_in_mins)
}
\arguments{
\item{exp_in_mins}{= (int) The number of minutes that the bearer token will be valid for.}
}
\value{
Update Bearer Token from secret key & returns working \code{environment} and saves updated tokens in 'rp_tokens.rds'
}
\description{
Get New Access/Bearer Token From Secret Key
}
\examples{
\dontrun{
  # Request New Bearer Token that expires in 120 minutes
  rp_getAccToken(exp_in_mins=120)
}
}
