% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_getAccts}
\alias{rp_getAccts}
\title{Get Public Account Info}
\usage{
rp_getAccts()
}
\value{
Returns a \code{data.frame} for the user's Public Brokerage Account
}
\description{
Get Public Account Info
}
\examples{
\dontrun{
  # Return Public Brokerage Account Information
    rp_getAccts()
}
}
