% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_getAcctsPort}
\alias{rp_getAcctsPort}
\title{Get Account Portfolio V2}
\usage{
rp_getAcctsPort(accountId)
}
\arguments{
\item{accountId}{= Public Brokerage Account Number}
}
\value{
Returns a \code{list} for the user's specific Public Brokerage Account
}
\description{
Get Account Portfolio V2
}
\examples{
\dontrun{
  # Return Public Brokerage Account Information
     my_acc <- rp_getAccts()
    my_port <- rp_getAcctsPort(accountId = my_acc$accountId)
}
}
