% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_getInstrument}
\alias{rp_getInstrument}
\title{Get Specific Instrument Information}
\usage{
rp_getInstrument(symbol, type)
}
\arguments{
\item{symbol}{= Trading Symbol Type: Ex. "AAPL"}

\item{type}{= Symbol Type Ex. "EQUITY", "OPTION", "MULTI_LEG_INSTRUMENT", "CRYPTO", "ALT", "TREASURY", "BOND", "INDEX"}
}
\value{
Retrieves specific trading instrument with optional filtering capabilities as a \code{data.frame}.
}
\description{
Get Specific Instrument Information
}
\examples{
\dontrun{
  # Fetches AAPL instrument trading information
    this_ins <- rp_getInstrument(symbol = "AAPL", type="EQUITY")
}
}
