% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_getOrderId}
\alias{rp_getOrderId}
\title{Order ID}
\usage{
rp_getOrderId()
}
\value{
An auto generated \code{character} string to use for placing orders
}
\description{
Order ID
}
\examples{
\dontrun{
  rp_getOrderId()
}
}
