% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPublic.R
\name{rp_hist_bars}
\alias{rp_hist_bars}
\title{Get bars v2 with aggregation}
\usage{
rp_hist_bars(
  symbol,
  period = "DAY",
  type = "EQUITY",
  aggregation,
  purchaseDate = NULL,
  include_ah = FALSE
)
}
\arguments{
\item{symbol}{= The symbol to fetch bar data for: Ex. "SPY"}

\item{period}{= one of DAY, WEEK, MONTH, QUARTER, HALF_YEAR, YEAR, FIVE_YEARS, YTD, SINCE_PURCHASE}

\item{type}{= Ticker Type: Ex. 'EQUITY','OPTION', 'CRYPTO', 'INDEX'}

\item{aggregation}{= Aggregation Ex.ONE_MINUTE, FIVE_MINUTES, TEN_MINUTES, FIFTEEN_MINUTES, THIRTY_MINUTES, ONE_HOUR, ONE_DAY, ONE_WEEK, ONE_MONTH, THREE_MONTHS,
SIX_MONTH, SONE_YEAR}

\item{purchaseDate}{= Required for SINCE_PURCHASE. YYYY-MM-DD}

\item{include_ah}{= Should after-hours data (if available) be included? Default is FALSE}
}
\value{
Fetch bar data for a given symbol and period and return as a \code{data.frame}.
}
\description{
Get bars v2 with aggregation
}
\examples{
\dontrun{
 # Fetches Bars for the DAY - include after-hours data
 df <- rp_hist_bars("TSLA", period = "DAY", type = "EQUITY", 
 aggregation = "FIVE_MINUTES", include_ah = TRUE)
 # Fetches Bars using SINCE_PURCHASE
 df <- rp_hist_bars("UBER", period = "MONTH", type = "EQUITY", 
 aggregation = "ONE_DAY", purchaseDate = "2026-04-01")
 # Fetches Bars for Options
 df <- rp_hist_bars("SPY260522P00739000", period = "DAY",
  type = "OPTION", aggregation = "ONE_MINUTE")
 # Fetches Bars for Crypto
 df <- rp_hist_bars("BTC", period = "WEEK", type = "CRYPTO",
  aggregation = "FIFTEEN_MINUTES")
}
}
